Checkpoint
I2C communication works (tried with a RPi as master) but definitely not complete. At some point, I feel either the specs or the way I'm using them might need to be revised.

Cascaded motor control loop is there but the HAL it links to for the DRV8304H motor drivers is currently empty. So, this code is from before I attempted the following integration.

I've been playing around with the TI EVM package for the DRV and have gotten a motor spinning but I haven't yet been able to successly extract just the functional parts of it so that I can package it up in the Driver_HAL.
I feel like I'm close though (I can extract enough to get the motor to twitch on command but not move in the way it does with the EVM package).
More to come soon.