Update Primer with more on Packed Bitfield Structs
It appears as though there's only one remaining issue to resolve: is there a way to guarantee if the bitfield containing struct will be loaded into memory or a register. The TI implementation is different depending on where the struct is stored but doesn't mention a way to be sure which.

Other than that, there may also be a question regarding if all types have to be of storage-unit integer type in the TI spec (some only allow char or int). I.e. is `uint16_t` allowed?