Fix lander <-> watchdog comms
still need to debug watchdog <-> hercules, but i suspect it is
a similar issue. there was some sort of race condition somewhere :/

Also, get rid of pbuf since it no longer is necessary. This also
lets us bump up the other buffers to 750 bytes, which should leave
more room for the hercules to send us stuff.